@@ -1,178 +1,114 @@
 {
-  "name": "remark",
-  "version": "4.2.2",
-  "description": "Markdown processor powered by plugins",
-  "license": "MIT",
-  "keywords": [
-    "markdown",
-    "markup",
-    "abstract",
-    "syntax",
-    "tree",
-    "json",
-    "ast",
-    "parse",
-    "stringify",
-    "bin",
-    "cli"
+  "_args": [
+    [
+      {
+        "raw": "remark@^6.0.1",
+        "scope": null,
+        "escapedName": "remark",
+        "name": "remark",
+        "rawSpec": "^6.0.1",
+        "spec": ">=6.0.1 <7.0.0",
+        "type": "range"
+      },
+      "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71836-1x3rf75/package.tgz"
+    ]
   ],
-  "dependencies": {
-    "camelcase": "^2.0.0",
-    "ccount": "^1.0.0",
-    "chalk": "^1.0.0",
-    "chokidar": "^1.0.5",
-    "collapse-white-space": "^1.0.0",
-    "commander": "^2.0.0",
-    "concat-stream": "^1.0.0",
-    "debug": "^2.0.0",
-    "elegant-spinner": "^1.0.0",
-    "extend": "^3.0.0",
-    "glob": "^7.0.0",
-    "globby": "^4.0.0",
-    "log-update": "^1.0.1",
-    "longest-streak": "^1.0.0",
-    "markdown-table": "^0.4.0",
-    "minimatch": "^3.0.0",
-    "npm-prefix": "^1.0.1",
-    "parse-entities": "^1.0.0",
-    "repeat-string": "^1.5.0",
-    "stringify-entities": "^1.0.0",
-    "to-vfile": "^1.0.0",
-    "trim": "^0.0.1",
-    "trim-trailing-lines": "^1.0.0",
-    "unified": "^3.0.0",
-    "unist-util-remove-position": "^1.0.0",
-    "user-home": "^2.0.0",
-    "vfile": "^1.1.0",
-    "vfile-find-down": "^1.0.0",
-    "vfile-find-up": "^1.0.0",
-    "vfile-location": "^2.0.0",
-    "vfile-reporter": "^1.5.0",
-    "ware": "^1.3.0"
+  "_from": "remark@>=6.0.1 <7.0.0",
+  "_id": "remark@6.0.1",
+  "_inCache": true,
+  "_installable": true,
+  "_location": "/atom-beautify/remark",
+  "_nodeVersion": "6.3.0",
+  "_npmOperationalInternal": {
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/remark-6.0.1.tgz_1472930278190_0.09580574021674693"
   },
-  "homepage": "http://remark.js.org",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wooorm/remark.git"
+  "_npmUser": {
+    "name": "wooorm",
+    "email": "tituswormer@gmail.com"
   },
-  "bugs": {
-    "url": "https://github.com/wooorm/remark/issues"
+  "_npmVersion": "3.10.6",
+  "_phantomChildren": {},
+  "_requested": {
+    "raw": "remark@^6.0.1",
+    "scope": null,
+    "escapedName": "remark",
+    "name": "remark",
+    "rawSpec": "^6.0.1",
+    "spec": ">=6.0.1 <7.0.0",
+    "type": "range"
   },
+  "_requiredBy": [
+    "/atom-beautify"
+  ],
+  "_resolved": "https://registry.npmjs.org/remark/-/remark-6.0.1.tgz",
+  "_shasum": "6bf09587a9436b76fc4c210582544bd8f8b02a94",
+  "_shrinkwrap": null,
+  "_spec": "remark@^6.0.1",
+  "_where": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71836-1x3rf75/package.tgz",
   "author": {
     "name": "Titus Wormer",
     "email": "tituswormer@gmail.com",
     "url": "http://wooorm.com"
+  },
+  "bin": {
+    "remark": "cli.js"
+  },
+  "bugs": {
+    "url": "https://github.com/wooorm/remark/issues"
   },
   "contributors": [
     {
       "name": "Titus Wormer",
       "email": "tituswormer@gmail.com",
       "url": "http://wooorm.com"
-    },
-    {
-      "name": "Eugene Sharygin",
-      "email": "eush77@gmail.com"
-    },
-    {
-      "name": "YJ Yang",
-      "email": "chcokr@gmail.com"
     }
   ],
-  "bin": {
-    "remark": "bin/remark"
+  "dependencies": {
+    "load-plugin": "^2.0.0",
+    "remark-parse": "^2.0.0",
+    "remark-stringify": "^2.0.0",
+    "unified": "^5.0.0"
   },
-  "directories": {
-    "man": "./man"
+  "description": "Markdown processor powered by plugins",
+  "devDependencies": {},
+  "directories": {},
+  "dist": {
+    "shasum": "6bf09587a9436b76fc4c210582544bd8f8b02a94",
+    "tarball": "https://registry.npmjs.org/remark/-/remark-6.0.1.tgz"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=0.11.0"
   },
   "files": [
     "index.js",
-    "lib",
-    "man",
-    "bin"
+    "cli.js"
   ],
-  "devDependencies": {
-    "bail": "^1.0.0",
-    "browserify": "^13.0.0",
-    "clone": "^1.0.1",
-    "escodegen": "^1.7.0",
-    "eslint": "^2.0.0",
-    "esmangle": "^1.0.0",
-    "esprima": "^2.6.0",
-    "istanbul": "^0.4.0",
-    "jscs": "^2.0.0",
-    "jscs-jsdoc": "^1.0.0",
-    "mocha": "^2.0.0",
-    "remark-comment-config": "^3.0.0",
-    "remark-github": "^4.0.1",
-    "remark-html": "^3.0.0",
-    "remark-license": "^2.0.0",
-    "remark-lint": "^3.0.0",
-    "remark-man": "^3.0.0",
-    "remark-retext": "^1.0.1",
-    "remark-toc": "^3.0.0",
-    "remark-usage": "^3.0.0",
-    "remark-validate-links": "^3.0.0",
-    "remark-yaml-config": "^3.0.0",
-    "retext": "^2.0.0",
-    "retext-equality": "^2.2.1",
-    "retext-profanities": "^1.1.0",
-    "retext-readability": "^1.0.0",
-    "retext-simplify": "^1.1.0"
-  },
-  "scripts": {
-    "build-man": "bin/remark doc/*.?.md --config-path .remarkrc-man --quiet",
-    "build-md": "bin/remark . --quiet --frail",
-    "build-version": "node script/build-version.js",
-    "build-options": "node script/build-options.js",
-    "build-bundle": "node script/build-bundle.js",
-    "lint-api": "eslint .",
-    "lint-style": "jscs --reporter inline .",
-    "lint": "npm run lint-api && npm run lint-style",
-    "build": "npm run build-version && npm run build-md && npm run build-options && npm run build-man && npm run build-bundle",
-    "test-api": "mocha --check-leaks test/index.js",
-    "test-api-extensive": "TEST_EXTENDED=true npm run test-api",
-    "test-cli": "bash test/cli.sh",
-    "test-coverage": "istanbul cover _mocha -- -- test/index.js",
-    "test-travis": "npm run build-md && npm run build-bundle && npm run lint && npm run test-coverage && npm run test-cli",
-    "test": "npm run build && npm run lint && npm run test-coverage && npm run test-cli"
-  },
-  "man": [
-    "/home/travis/build/wooorm/remark/man/remark.1",
-    "/home/travis/build/wooorm/remark/man/remark.3",
-    "/home/travis/build/wooorm/remark/man/remarkconfig.7",
-    "/home/travis/build/wooorm/remark/man/remarkignore.5",
-    "/home/travis/build/wooorm/remark/man/remarkplugin.3",
-    "/home/travis/build/wooorm/remark/man/remarkplugin.7",
-    "/home/travis/build/wooorm/remark/man/remarkrc.5",
-    "/home/travis/build/wooorm/remark/man/remarksetting.7"
+  "homepage": "http://remark.js.org",
+  "keywords": [
+    "markdown",
+    "abstract",
+    "syntax",
+    "tree",
+    "ast",
+    "parse",
+    "stringify",
+    "process"
   ],
-  "gitHead": "fc5e502cfc908db401394337548131cc739e777f",
-  "_id": "remark@4.2.2",
-  "_shasum": "42dc5d0b3a6a2d5443e7cbdbb2a3202854be698f",
-  "_from": "remark@>=4.2.2 <5.0.0",
-  "_npmVersion": "3.3.6",
-  "_nodeVersion": "5.0.0",
-  "_npmUser": {
-    "name": "wooorm",
-    "email": "tituswormer@gmail.com"
-  },
-  "dist": {
-    "shasum": "42dc5d0b3a6a2d5443e7cbdbb2a3202854be698f",
-    "tarball": "https://registry.npmjs.org/remark/-/remark-4.2.2.tgz"
-  },
+  "license": "MIT",
   "maintainers": [
     {
       "name": "wooorm",
       "email": "tituswormer@gmail.com"
     }
   ],
-  "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/remark-4.2.2.tgz_1461321251732_0.6673849069047719"
+  "name": "remark",
+  "optionalDependencies": {},
+  "readme": "ERROR: No README data found!",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/wooorm/remark/tree/master/packages/remark"
   },
-  "_resolved": "https://registry.npmjs.org/remark/-/remark-4.2.2.tgz",
-  "readme": "ERROR: No README data found!"
+  "scripts": {},
+  "version": "6.0.1"
 }
