@@ -1,39 +1,75 @@
 {
-  "name": "linter-csslint",
-  "keywords": [
-    "csslint",
-    "css",
-    "linter",
-    "lint"
+  "_args": [
+    [
+      {
+        "raw": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+        "scope": null,
+        "escapedName": null,
+        "name": null,
+        "rawSpec": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+        "spec": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+        "type": "local"
+      },
+      "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/apm-install-dir-11694-71856-1kian67"
+    ]
   ],
-  "main": "./lib/main",
-  "version": "1.3.2",
-  "description": "Lint CSS on the fly, using csslint",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/AtomLinter/linter-csslint.git"
+  "_from": "../d-11694-71856-8s8j9h/package.tgz",
+  "_id": "linter-csslint@1.3.3",
+  "_inCache": true,
+  "_installable": true,
+  "_location": "/linter-csslint",
+  "_phantomChildren": {},
+  "_requested": {
+    "raw": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+    "scope": null,
+    "escapedName": null,
+    "name": null,
+    "rawSpec": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+    "spec": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+    "type": "local"
   },
-  "license": "MIT",
+  "_requiredBy": [
+    "#USER"
+  ],
+  "_resolved": "file:../d-11694-71856-8s8j9h/package.tgz",
+  "_shasum": "15a49e57d325ac8ecc1f00aabaf78c85b7b2d57c",
+  "_shrinkwrap": null,
+  "_spec": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/d-11694-71856-8s8j9h/package.tgz",
+  "_where": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/apm-install-dir-11694-71856-1kian67",
+  "bugs": {
+    "url": "https://github.com/AtomLinter/linter-csslint/issues"
+  },
+  "configSchema": {
+    "disableTimeout": {
+      "type": "boolean",
+      "description": "Disable the 10 second execution timeout",
+      "default": false
+    }
+  },
   "dependencies": {
-    "atomlinter-csslint": "0.10.0",
-    "atom-linter": "^4.3.0",
-    "atom-package-deps": "^4.0.1"
+    "atom-linter": "^8.0.0",
+    "atom-package-deps": "^4.0.1",
+    "atomlinter-csslint": "0.10.0"
   },
+  "description": "Lint CSS on the fly, using csslint",
   "devDependencies": {
-    "coffeelint": "^1.14.1",
-    "eslint": "^2.2.0",
-    "babel-eslint": "^5.0.0",
-    "eslint-config-airbnb": "^6.0.2"
-  },
-  "package-deps": [
-    "linter"
-  ],
-  "scripts": {
-    "lint": "coffeelint lib"
+    "eslint": "^3.6.1",
+    "eslint-config-airbnb-base": "^8.0.0",
+    "eslint-plugin-import": "^1.16.0"
   },
   "eslintConfig": {
-    "extends": "airbnb/base",
-    "parser": "babel-eslint",
+    "extends": "airbnb-base",
+    "rules": {
+      "global-require": "off",
+      "import/no-unresolved": [
+        "error",
+        {
+          "ignore": [
+            "atom"
+          ]
+        }
+      ]
+    },
     "globals": {
       "atom": true
     },
