@@ -10,24 +10,24 @@
         "spec": ">=1.0.7 <2.0.0",
         "type": "range"
       },
-      "/private/var/folders/_p/49n358xd23578jf2724w2w5r0000gn/T/apm-install-dir-11688-15360-uae8c5/node_modules/linter-eslint/node_modules/js-yaml"
+      "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/apm-install-dir-11694-71866-4vxgcr/node_modules/linter-eslint/node_modules/js-yaml"
     ]
   ],
   "_from": "argparse@>=1.0.7 <2.0.0",
-  "_id": "argparse@1.0.7",
+  "_id": "argparse@1.0.9",
   "_inCache": true,
   "_installable": true,
   "_location": "/linter-eslint/argparse",
-  "_nodeVersion": "4.3.0",
+  "_nodeVersion": "6.5.0",
   "_npmOperationalInternal": {
     "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/argparse-1.0.7.tgz_1458226887302_0.8524546672124416"
+    "tmp": "tmp/argparse-1.0.9.tgz_1475177461025_0.33920647646300495"
   },
   "_npmUser": {
     "name": "vitaly",
     "email": "vitaly@rcdesign.ru"
   },
-  "_npmVersion": "2.14.12",
+  "_npmVersion": "3.10.3",
   "_phantomChildren": {},
   "_requested": {
     "raw": "argparse@^1.0.7",
@@ -41,11 +41,11 @@
   "_requiredBy": [
     "/linter-eslint/js-yaml"
   ],
-  "_resolved": "https://registry.npmjs.org/argparse/-/argparse-1.0.7.tgz",
-  "_shasum": "c289506480557810f14a8bc62d7a06f63ed7f951",
+  "_resolved": "https://registry.npmjs.org/argparse/-/argparse-1.0.9.tgz",
+  "_shasum": "73d83bc263f86e97f8cc4f6bae1b0e90a7d22c86",
   "_shrinkwrap": null,
   "_spec": "argparse@^1.0.7",
-  "_where": "/private/var/folders/_p/49n358xd23578jf2724w2w5r0000gn/T/apm-install-dir-11688-15360-uae8c5/node_modules/linter-eslint/node_modules/js-yaml",
+  "_where": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/apm-install-dir-11694-71866-4vxgcr/node_modules/linter-eslint/node_modules/js-yaml",
   "bugs": {
     "url": "https://github.com/nodeca/argparse/issues"
   },
@@ -62,22 +62,22 @@
   },
   "description": "Very powerful CLI arguments parser. Native port of argparse - python's options parsing library",
   "devDependencies": {
-    "eslint": "2.0.0-rc.0",
-    "eslint-plugin-nodeca": "~1.0.3",
-    "mocha": "*",
-    "ndoc": "^3.1.0"
+    "eslint": "^2.13.1",
+    "istanbul": "^0.4.5",
+    "mocha": "^3.1.0",
+    "ndoc": "^5.0.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "c289506480557810f14a8bc62d7a06f63ed7f951",
-    "tarball": "https://registry.npmjs.org/argparse/-/argparse-1.0.7.tgz"
+    "shasum": "73d83bc263f86e97f8cc4f6bae1b0e90a7d22c86",
+    "tarball": "https://registry.npmjs.org/argparse/-/argparse-1.0.9.tgz"
   },
   "files": [
     "index.js",
     "lib/"
   ],
-  "gitHead": "2d243d3062affb01f49f5f00e8ca8464a15e9374",
-  "homepage": "https://github.com/nodeca/argparse",
+  "gitHead": "acb39f2d726b90d2eadf9e6574a938d6250ad248",
+  "homepage": "https://github.com/nodeca/argparse#readme",
   "keywords": [
     "cli",
     "parser",
@@ -102,5 +102,5 @@
   "scripts": {
     "test": "make test"
   },
-  "version": "1.0.7"
+  "version": "1.0.9"
 }
