@@ -2,46 +2,50 @@
   "_args": [
     [
       {
-        "raw": "jsonpointer@2.0.0",
+        "raw": "jsonpointer@^4.0.0",
         "scope": null,
         "escapedName": "jsonpointer",
         "name": "jsonpointer",
-        "rawSpec": "2.0.0",
-        "spec": "2.0.0",
-        "type": "version"
+        "rawSpec": "^4.0.0",
+        "spec": ">=4.0.0 <5.0.0",
+        "type": "range"
       },
-      "/private/var/folders/_p/49n358xd23578jf2724w2w5r0000gn/T/apm-install-dir-11688-15360-uae8c5/node_modules/linter-eslint/node_modules/is-my-json-valid"
+      "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/apm-install-dir-11694-71866-4vxgcr/node_modules/linter-eslint/node_modules/is-my-json-valid"
     ]
   ],
-  "_from": "jsonpointer@2.0.0",
-  "_id": "jsonpointer@2.0.0",
+  "_from": "jsonpointer@>=4.0.0 <5.0.0",
+  "_id": "jsonpointer@4.0.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/linter-eslint/jsonpointer",
-  "_nodeVersion": "0.10.36",
+  "_nodeVersion": "6.1.0",
+  "_npmOperationalInternal": {
+    "host": "packages-16-east.internal.npmjs.com",
+    "tmp": "tmp/jsonpointer-4.0.0.tgz_1463651460494_0.02921536797657609"
+  },
   "_npmUser": {
     "name": "marcbachmann",
     "email": "marc.brookman@gmail.com"
   },
-  "_npmVersion": "2.10.1",
+  "_npmVersion": "3.8.6",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "jsonpointer@2.0.0",
+    "raw": "jsonpointer@^4.0.0",
     "scope": null,
     "escapedName": "jsonpointer",
     "name": "jsonpointer",
-    "rawSpec": "2.0.0",
-    "spec": "2.0.0",
-    "type": "version"
+    "rawSpec": "^4.0.0",
+    "spec": ">=4.0.0 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
     "/linter-eslint/is-my-json-valid"
   ],
-  "_resolved": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-2.0.0.tgz",
-  "_shasum": "3af1dd20fe85463910d469a385e33017d2a030d9",
+  "_resolved": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-4.0.0.tgz",
+  "_shasum": "6661e161d2fc445f19f98430231343722e1fcbd5",
   "_shrinkwrap": null,
-  "_spec": "jsonpointer@2.0.0",
-  "_where": "/private/var/folders/_p/49n358xd23578jf2724w2w5r0000gn/T/apm-install-dir-11688-15360-uae8c5/node_modules/linter-eslint/node_modules/is-my-json-valid",
+  "_spec": "jsonpointer@^4.0.0",
+  "_where": "/private/var/folders/ct/n7hjn86j5m34yygr29bgdz0r0000gp/T/apm-install-dir-11694-71866-4vxgcr/node_modules/linter-eslint/node_modules/is-my-json-valid",
   "author": {
     "name": "Jan Lehnardt",
     "email": "jan@apache.org"
@@ -53,20 +57,26 @@
     {
       "name": "Joe Hildebrand",
       "email": "joe-github@cursive.net"
+    },
+    {
+      "name": "Marc Bachmann",
+      "email": "marc.brookman@gmail.com"
     }
   ],
   "dependencies": {},
   "description": "Simple JSON Addressing.",
-  "devDependencies": {},
+  "devDependencies": {
+    "standard": "^5.3.1"
+  },
   "directories": {},
   "dist": {
-    "shasum": "3af1dd20fe85463910d469a385e33017d2a030d9",
-    "tarball": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-2.0.0.tgz"
+    "shasum": "6661e161d2fc445f19f98430231343722e1fcbd5",
+    "tarball": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-4.0.0.tgz"
   },
   "engines": {
-    "node": ">=0.6.0"
+    "node": ">=0.10.0"
   },
-  "gitHead": "26ea4a5c0fcb6d9a2e87f733403791dd05637af8",
+  "gitHead": "2d46030ba6df41b566934c7202e31fb65058de71",
   "homepage": "https://github.com/janl/node-jsonpointer#readme",
   "license": "MIT",
   "main": "./jsonpointer",
@@ -88,7 +98,7 @@
     "url": "git+ssh://git@github.com/janl/node-jsonpointer.git"
   },
   "scripts": {
-    "test": "node test.js"
+    "test": "standard && node test.js"
   },
   "tags": [
     "util",
@@ -96,5 +106,5 @@
     "util",
     "utility"
   ],
-  "version": "2.0.0"
+  "version": "4.0.0"
 }
