@@ -1,70 +1,78 @@
 'use babel';
 
-export default class ProjectManager {
-  static activate() {
-    const CompositeDisposable = require('atom').CompositeDisposable;
-    this.disposables = new CompositeDisposable();
-    this.projects = require('./projects');
+import { autorun } from 'mobx';
+import { CompositeDisposable } from 'atom';
+import manager from './Manager';
+import { SAVE_URI, EDIT_URI } from './views/view-uri';
 
-    this.disposables.add(atom.commands.add('atom-workspace', {
-      'project-manager:list-projects': () => {
-        if (!this.projectsListView) {
-          const ProjectsListView = require('./projects-list-view');
-          this.projectsListView = new ProjectsListView();
-        }
+let disposables = null;
+let projectsListView = null;
+let FileStore = null;
 
-        this.projectsListView.toggle();
-      },
+export function editComponent() {
+  const EditView = require('./views/EditView');
 
-      'project-manager:save-project': () => {
-        if (!this.saveDialog) {
-          const SaveDialog = require('./save-dialog');
-          this.saveDialog = new SaveDialog();
-        }
+  return new EditView({ project: manager.activeProject });
+}
 
-        this.saveDialog.attach();
-      },
+export function activate() {
+  disposables = new CompositeDisposable();
 
-      'project-manager:edit-projects': () => {
-        if (!this.db) {
-          this.db = require('./db');
-        }
+  disposables.add(atom.workspace.addOpener((uri) => {
+    if (uri === EDIT_URI) {
+      return editComponent();
+    }
 
-        atom.workspace.open(this.db.file());
+    return null;
+  }));
+
+  disposables.add(atom.commands.add('atom-workspace', {
+    'project-manager:list-projects': () => {
+      if (!this.projectsListView) {
+        const ProjectsListView = require('./views/projects-list-view');
+
+        projectsListView = new ProjectsListView();
       }
-    }));
 
-    atom.project.onDidChangePaths(() => this.updatePaths());
-    this.loadProject();
-  }
-
-  static loadProject() {
-    this.projects.getCurrent(project => {
-      if (project) {
-        this.project = project;
-        this.project.load();
+      projectsListView.toggle();
+    },
+    'project-manager:edit-projects': () => {
+      if (!FileStore) {
+        FileStore = require('./stores/FileStore');
       }
-    });
-  }
 
-  static updatePaths() {
-    this.projects.getCurrent(project => {
-      const newPaths = atom.project.getPaths();
-      const currentRoot = newPaths.length ? newPaths[0] : null;
+      atom.workspace.open(FileStore.getPath());
+    },
+    'project-manager:save-project': () => {
+      atom.workspace.open(SAVE_URI);
+    },
+    'project-manager:edit-project': () => {
+      atom.workspace.open(EDIT_URI);
+    },
+  }));
+}
 
-      if (project.rootPath === currentRoot) {
-        project.set('paths', newPaths);
-      }
-    });
-  }
+export function deactivate() {
+  disposables.dispose();
+}
 
-  static provideProjects() {
-    return {
-      projects: this.projects
-    };
-  }
-
-  static deactivate() {
-    this.disposables.dispose();
-  }
+export function provideProjects() {
+  return {
+    getProjects: (callback) => {
+      autorun(() => {
+        callback(manager.projects);
+      });
+    },
+    getProject: (callback) => {
+      autorun(() => {
+        callback(manager.activeProject);
+      });
+    },
+    saveProject: (project) => {
+      manager.saveProject(project);
+    },
+    openProject: (project) => {
+      manager.open(project);
+    },
+  };
 }
