@@ -1,1 +1,1 @@
-window.searchData = [{"t":"Autopep8","p":"class/Autopep8.html"},{"t":"Beautifier","p":"class/Beautifier.html"},{"t":"ClangFormat","p":"class/ClangFormat.html"},{"t":"CoffeeFmt","p":"class/CoffeeFmt.html"},{"t":"CoffeeFormatter","p":"class/CoffeeFormatter.html"},{"t":"Crystal","p":"class/Crystal.html"},{"t":"JSBeautify","p":"class/JSBeautify.html"},{"t":"Dfmt","p":"class/Dfmt.html"},{"t":"ElmFormat","p":"class/ElmFormat.html"},{"t":"ErlTidy","p":"class/ErlTidy.html"},{"t":"FortranBeautifier","p":"class/FortranBeautifier.html"},{"t":"Gherkin","p":"class/Gherkin.html"},{"t":"Gofmt","p":"class/Gofmt.html"},{"t":"HTMLBeautifier","p":"class/HTMLBeautifier.html"},{"t":"Beautifiers","p":"class/Beautifiers.html"},{"t":"JSBeautify","p":"class/JSBeautify.html"},{"t":"JSCSFixer","p":"class/JSCSFixer.html"},{"t":"LatexBeautify","p":"class/LatexBeautify.html"},{"t":"Lua","p":"class/Lua.html"},{"t":"MarkoBeautifier","p":"class/MarkoBeautifier.html"},{"t":"OCPIndent","p":"class/OCPIndent.html"},{"t":"PerlTidy","p":"class/PerlTidy.html"},{"t":"PHPCSFixer","p":"class/PHPCSFixer.html"},{"t":"PHPCBF","p":"class/PHPCBF.html"},{"t":"PrettyDiff","p":"class/PrettyDiff.html"},{"t":"PugBeautify","p":"class/PugBeautify.html"},{"t":"PuppetFix","p":"class/PuppetFix.html"},{"t":"Remark","p":"class/Remark.html"},{"t":"Rubocop","p":"class/Rubocop.html"},{"t":"RubyBeautify","p":"class/RubyBeautify.html"},{"t":"Rustfmt","p":"class/Rustfmt.html"},{"t":"Sqlformat","p":"class/Sqlformat.html"},{"t":"StylishHaskell","p":"class/StylishHaskell.html"},{"t":"TidyMarkdown","p":"class/TidyMarkdown.html"},{"t":"TypeScriptFormatter","p":"class/TypeScriptFormatter.html"},{"t":"Uncrustify","p":"class/Uncrustify.html"},{"t":"Yapf","p":"class/Yapf.html"},{"t":"LoadingView","p":"class/LoadingView.html"},{"t":"MessageView","p":"class/MessageView.html"},{"t":"src/beautifiers/autopep8.coffee","p":"file/src/beautifiers/autopep8.coffee.html"},{"t":"src/beautifiers/beautifier.coffee","p":"file/src/beautifiers/beautifier.coffee.html"},{"t":"src/beautifiers/clang-format.coffee","p":"file/src/beautifiers/clang-format.coffee.html"},{"t":"src/beautifiers/coffee-fmt.coffee","p":"file/src/beautifiers/coffee-fmt.coffee.html"},{"t":"src/beautifiers/coffee-formatter.coffee","p":"file/src/beautifiers/coffee-formatter.coffee.html"},{"t":"src/beautifiers/crystal.coffee","p":"file/src/beautifiers/crystal.coffee.html"},{"t":"src/beautifiers/csscomb.coffee","p":"file/src/beautifiers/csscomb.coffee.html"},{"t":"src/beautifiers/dfmt.coffee","p":"file/src/beautifiers/dfmt.coffee.html"},{"t":"src/beautifiers/elm-format.coffee","p":"file/src/beautifiers/elm-format.coffee.html"},{"t":"src/beautifiers/erl_tidy.coffee","p":"file/src/beautifiers/erl_tidy.coffee.html"},{"t":"src/beautifiers/fortran-beautifier/index.coffee","p":"file/src/beautifiers/fortran-beautifier/index.coffee.html"},{"t":"src/beautifiers/gherkin.coffee","p":"file/src/beautifiers/gherkin.coffee.html"},{"t":"src/beautifiers/gofmt.coffee","p":"file/src/beautifiers/gofmt.coffee.html"},{"t":"src/beautifiers/htmlbeautifier.coffee","p":"file/src/beautifiers/htmlbeautifier.coffee.html"},{"t":"src/beautifiers/index.coffee","p":"file/src/beautifiers/index.coffee.html"},{"t":"src/beautifiers/js-beautify.coffee","p":"file/src/beautifiers/js-beautify.coffee.html"},{"t":"src/beautifiers/jscs.coffee","p":"file/src/beautifiers/jscs.coffee.html"},{"t":"src/beautifiers/latex-beautify.coffee","p":"file/src/beautifiers/latex-beautify.coffee.html"},{"t":"src/beautifiers/lua-beautifier/index.coffee","p":"file/src/beautifiers/lua-beautifier/index.coffee.html"},{"t":"src/beautifiers/marko-beautifier.coffee","p":"file/src/beautifiers/marko-beautifier.coffee.html"},{"t":"src/beautifiers/ocp-indent.coffee","p":"file/src/beautifiers/ocp-indent.coffee.html"},{"t":"src/beautifiers/perltidy.coffee","p":"file/src/beautifiers/perltidy.coffee.html"},{"t":"src/beautifiers/php-cs-fixer.coffee","p":"file/src/beautifiers/php-cs-fixer.coffee.html"},{"t":"src/beautifiers/phpcbf.coffee","p":"file/src/beautifiers/phpcbf.coffee.html"},{"t":"src/beautifiers/prettydiff.coffee","p":"file/src/beautifiers/prettydiff.coffee.html"},{"t":"src/beautifiers/pug-beautify.coffee","p":"file/src/beautifiers/pug-beautify.coffee.html"},{"t":"src/beautifiers/puppet-fix.coffee","p":"file/src/beautifiers/puppet-fix.coffee.html"},{"t":"src/beautifiers/remark.coffee","p":"file/src/beautifiers/remark.coffee.html"},{"t":"src/beautifiers/rubocop.coffee","p":"file/src/beautifiers/rubocop.coffee.html"},{"t":"src/beautifiers/ruby-beautify.coffee","p":"file/src/beautifiers/ruby-beautify.coffee.html"},{"t":"src/beautifiers/rustfmt.coffee","p":"file/src/beautifiers/rustfmt.coffee.html"},{"t":"src/beautifiers/sqlformat.coffee","p":"file/src/beautifiers/sqlformat.coffee.html"},{"t":"src/beautifiers/stylish-haskell.coffee","p":"file/src/beautifiers/stylish-haskell.coffee.html"},{"t":"src/beautifiers/tidy-markdown.coffee","p":"file/src/beautifiers/tidy-markdown.coffee.html"},{"t":"src/beautifiers/typescript-formatter.coffee","p":"file/src/beautifiers/typescript-formatter.coffee.html"},{"t":"src/beautifiers/uncrustify/cfg.coffee","p":"file/src/beautifiers/uncrustify/cfg.coffee.html"},{"t":"src/beautifiers/uncrustify/index.coffee","p":"file/src/beautifiers/uncrustify/index.coffee.html"},{"t":"src/beautifiers/yapf.coffee","p":"file/src/beautifiers/yapf.coffee.html"},{"t":"src/beautify.coffee","p":"file/src/beautify.coffee.html"},{"t":"src/config.coffee","p":"file/src/config.coffee.html"},{"t":"src/languages/apex.coffee","p":"file/src/languages/apex.coffee.html"},{"t":"src/languages/arduino.coffee","p":"file/src/languages/arduino.coffee.html"},{"t":"src/languages/c-sharp.coffee","p":"file/src/languages/c-sharp.coffee.html"},{"t":"src/languages/c.coffee","p":"file/src/languages/c.coffee.html"},{"t":"src/languages/coffeescript.coffee","p":"file/src/languages/coffeescript.coffee.html"},{"t":"src/languages/coldfusion.coffee","p":"file/src/languages/coldfusion.coffee.html"},{"t":"src/languages/cpp.coffee","p":"file/src/languages/cpp.coffee.html"},{"t":"src/languages/crystal.coffee","p":"file/src/languages/crystal.coffee.html"},{"t":"src/languages/css.coffee","p":"file/src/languages/css.coffee.html"},{"t":"src/languages/csv.coffee","p":"file/src/languages/csv.coffee.html"},{"t":"src/languages/d.coffee","p":"file/src/languages/d.coffee.html"},{"t":"src/languages/ejs.coffee","p":"file/src/languages/ejs.coffee.html"},{"t":"src/languages/elm.coffee","p":"file/src/languages/elm.coffee.html"},{"t":"src/languages/erb.coffee","p":"file/src/languages/erb.coffee.html"},{"t":"src/languages/erlang.coffee","p":"file/src/languages/erlang.coffee.html"},{"t":"src/languages/fortran.coffee","p":"file/src/languages/fortran.coffee.html"},{"t":"src/languages/gherkin.coffee","p":"file/src/languages/gherkin.coffee.html"},{"t":"src/languages/go.coffee","p":"file/src/languages/go.coffee.html"},{"t":"src/languages/handlebars.coffee","p":"file/src/languages/handlebars.coffee.html"},{"t":"src/languages/haskell.coffee","p":"file/src/languages/haskell.coffee.html"},{"t":"src/languages/html.coffee","p":"file/src/languages/html.coffee.html"},{"t":"src/languages/index.coffee","p":"file/src/languages/index.coffee.html"},{"t":"src/languages/jade.coffee","p":"file/src/languages/jade.coffee.html"},{"t":"src/languages/java.coffee","p":"file/src/languages/java.coffee.html"},{"t":"src/languages/javascript.coffee","p":"file/src/languages/javascript.coffee.html"},{"t":"src/languages/json.coffee","p":"file/src/languages/json.coffee.html"},{"t":"src/languages/jsx.coffee","p":"file/src/languages/jsx.coffee.html"},{"t":"src/languages/latex.coffee","p":"file/src/languages/latex.coffee.html"},{"t":"src/languages/less.coffee","p":"file/src/languages/less.coffee.html"},{"t":"src/languages/lua.coffee","p":"file/src/languages/lua.coffee.html"},{"t":"src/languages/markdown.coffee","p":"file/src/languages/markdown.coffee.html"},{"t":"src/languages/marko.coffee","p":"file/src/languages/marko.coffee.html"},{"t":"src/languages/mustache.coffee","p":"file/src/languages/mustache.coffee.html"},{"t":"src/languages/objective-c.coffee","p":"file/src/languages/objective-c.coffee.html"},{"t":"src/languages/ocaml.coffee","p":"file/src/languages/ocaml.coffee.html"},{"t":"src/languages/pawn.coffee","p":"file/src/languages/pawn.coffee.html"},{"t":"src/languages/perl.coffee","p":"file/src/languages/perl.coffee.html"},{"t":"src/languages/php.coffee","p":"file/src/languages/php.coffee.html"},{"t":"src/languages/puppet.coffee","p":"file/src/languages/puppet.coffee.html"},{"t":"src/languages/python.coffee","p":"file/src/languages/python.coffee.html"},{"t":"src/languages/riotjs.coffee","p":"file/src/languages/riotjs.coffee.html"},{"t":"src/languages/ruby.coffee","p":"file/src/languages/ruby.coffee.html"},{"t":"src/languages/rust.coffee","p":"file/src/languages/rust.coffee.html"},{"t":"src/languages/sass.coffee","p":"file/src/languages/sass.coffee.html"},{"t":"src/languages/scss.coffee","p":"file/src/languages/scss.coffee.html"},{"t":"src/languages/spacebars.coffee","p":"file/src/languages/spacebars.coffee.html"},{"t":"src/languages/sql.coffee","p":"file/src/languages/sql.coffee.html"},{"t":"src/languages/svg.coffee","p":"file/src/languages/svg.coffee.html"},{"t":"src/languages/swig.coffee","p":"file/src/languages/swig.coffee.html"},{"t":"src/languages/tss.coffee","p":"file/src/languages/tss.coffee.html"},{"t":"src/languages/twig.coffee","p":"file/src/languages/twig.coffee.html"},{"t":"src/languages/typescript.coffee","p":"file/src/languages/typescript.coffee.html"},{"t":"src/languages/vala.coffee","p":"file/src/languages/vala.coffee.html"},{"t":"src/languages/visualforce.coffee","p":"file/src/languages/visualforce.coffee.html"},{"t":"src/languages/xml.coffee","p":"file/src/languages/xml.coffee.html"},{"t":"src/languages/xtemplate.coffee","p":"file/src/languages/xtemplate.coffee.html"},{"t":"src/views/loading-view.coffee","p":"file/src/views/loading-view.coffee.html"},{"t":"src/views/message-view.coffee","p":"file/src/views/message-view.coffee.html"},{"t":"README.md","p":"extra/README.md.html"},{"t":"CHANGELOG.md","p":"extra/CHANGELOG.md.html"},{"t":"MessageView#addMessage","p":"class/MessageView.html#addMessage-dynamic"},{"t":"ErlTidy#beautify","p":"class/ErlTidy.html#beautify-dynamic"},{"t":"JSCSFixer#beautify","p":"class/JSCSFixer.html#beautify-dynamic"},{"t":"src/beautify.coffee~beautify","p":"file/src/beautify.coffee.html#beautify-"},{"t":"JSBeautify#beautify","p":"class/JSBeautify.html#beautify-dynamic"},{"t":"Rubocop#beautify","p":"class/Rubocop.html#beautify-dynamic"},{"t":"RubyBeautify#beautify","p":"class/RubyBeautify.html#beautify-dynamic"},{"t":"LatexBeautify#beautify","p":"class/LatexBeautify.html#beautify-dynamic"},{"t":"Autopep8#beautify","p":"class/Autopep8.html#beautify-dynamic"},{"t":"Lua#beautify","p":"class/Lua.html#beautify-dynamic"},{"t":"Remark#beautify","p":"class/Remark.html#beautify-dynamic"},{"t":"MarkoBeautifier#beautify","p":"class/MarkoBeautifier.html#beautify-dynamic"},{"t":"PuppetFix#beautify","p":"class/PuppetFix.html#beautify-dynamic"},{"t":"OCPIndent#beautify","p":"class/OCPIndent.html#beautify-dynamic"},{"t":"Rustfmt#beautify","p":"class/Rustfmt.html#beautify-dynamic"},{"t":"Sqlformat#beautify","p":"class/Sqlformat.html#beautify-dynamic"},{"t":"Beautifiers#beautify","p":"class/Beautifiers.html#beautify-dynamic"},{"t":"StylishHaskell#beautify","p":"class/StylishHaskell.html#beautify-dynamic"},{"t":"Yapf#beautify","p":"class/Yapf.html#beautify-dynamic"},{"t":"TidyMarkdown#beautify","p":"class/TidyMarkdown.html#beautify-dynamic"},{"t":"TypeScriptFormatter#beautify","p":"class/TypeScriptFormatter.html#beautify-dynamic"},{"t":"ClangFormat#beautify","p":"class/ClangFormat.html#beautify-dynamic"},{"t":"CoffeeFmt#beautify","p":"class/CoffeeFmt.html#beautify-dynamic"},{"t":"CoffeeFormatter#beautify","p":"class/CoffeeFormatter.html#beautify-dynamic"},{"t":"Crystal#beautify","p":"class/Crystal.html#beautify-dynamic"},{"t":"JSBeautify#beautify","p":"class/JSBeautify.html#beautify-dynamic"},{"t":"Dfmt#beautify","p":"class/Dfmt.html#beautify-dynamic"},{"t":"ElmFormat#beautify","p":"class/ElmFormat.html#beautify-dynamic"},{"t":"Uncrustify#beautify","p":"class/Uncrustify.html#beautify-dynamic"},{"t":"FortranBeautifier#beautify","p":"class/FortranBeautifier.html#beautify-dynamic"},{"t":"Gherkin#beautify","p":"class/Gherkin.html#beautify-dynamic"},{"t":"Gofmt#beautify","p":"class/Gofmt.html#beautify-dynamic"},{"t":"HTMLBeautifier#beautify","p":"class/HTMLBeautifier.html#beautify-dynamic"},{"t":"PerlTidy#beautify","p":"class/PerlTidy.html#beautify-dynamic"},{"t":"PugBeautify#beautify","p":"class/PugBeautify.html#beautify-dynamic"},{"t":"PrettyDiff#beautify","p":"class/PrettyDiff.html#beautify-dynamic"},{"t":"PHPCBF#beautify","p":"class/PHPCBF.html#beautify-dynamic"},{"t":"PHPCSFixer#beautify","p":"class/PHPCSFixer.html#beautify-dynamic"},{"t":"src/beautify.coffee~beautifyDirectory","p":"file/src/beautify.coffee.html#beautifyDirectory-"},{"t":"src/beautify.coffee~beautifyFile","p":"file/src/beautify.coffee.html#beautifyFile-"},{"t":"src/beautify.coffee~beautifyFilePath","p":"file/src/beautify.coffee.html#beautifyFilePath-"},{"t":"LatexBeautify#buildConfigFile","p":"class/LatexBeautify.html#buildConfigFile-dynamic"},{"t":"MessageView#clearMessages","p":"class/MessageView.html#clearMessages-dynamic"},{"t":"PuppetFix#cli","p":"class/PuppetFix.html#cli-dynamic"},{"t":"PerlTidy#cli","p":"class/PerlTidy.html#cli-dynamic"},{"t":"MessageView#close","p":"class/MessageView.html#close-dynamic"},{"t":"Beautifier#commandNotFoundError","p":"class/Beautifier.html#commandNotFoundError-dynamic"},{"t":"MessageView#constructor","p":"class/MessageView.html#constructor-dynamic"},{"t":"Beautifier#constructor","p":"class/Beautifier.html#constructor-dynamic"},{"t":"Beautifiers#constructor","p":"class/Beautifiers.html#constructor-dynamic"},{"t":"MessageView.content","p":"class/MessageView.html#content-static"},{"t":"LoadingView.content","p":"class/LoadingView.html#content-static"},{"t":"src/beautify.coffee~debug","p":"file/src/beautify.coffee.html#debug-"},{"t":"Beautifier#deprecate","p":"class/Beautifier.html#deprecate-dynamic"},{"t":"MessageView#destroy","p":"class/MessageView.html#destroy-dynamic"},{"t":"ClangFormat#dumpToFile","p":"class/ClangFormat.html#dumpToFile-dynamic"},{"t":"Beautifiers#encodeURI","p":"class/Beautifiers.html#encodeURI-dynamic"},{"t":"Beautifiers#findConfig","p":"class/Beautifiers.html#findConfig-dynamic"},{"t":"Beautifiers#findFile","p":"class/Beautifiers.html#findFile-dynamic"},{"t":"Beautifier#findFile","p":"class/Beautifier.html#findFile-dynamic"},{"t":"Beautifiers#getBeautifierForLanguage","p":"class/Beautifiers.html#getBeautifierForLanguage-dynamic"},{"t":"Beautifiers#getBeautifiers","p":"class/Beautifiers.html#getBeautifiers-dynamic"},{"t":"Beautifiers#getConfig","p":"class/Beautifiers.html#getConfig-dynamic"},{"t":"Beautifiers#getConfigOptionsFromSettings","p":"class/Beautifiers.html#getConfigOptionsFromSettings-dynamic"},{"t":"src/beautify.coffee~getCursors","p":"file/src/beautify.coffee.html#getCursors-"},{"t":"Beautifiers#getLanguage","p":"class/Beautifiers.html#getLanguage-dynamic"},{"t":"Beautifiers#getOptions","p":"class/Beautifiers.html#getOptions-dynamic"},{"t":"Beautifiers#getOptionsForLanguage","p":"class/Beautifiers.html#getOptionsForLanguage-dynamic"},{"t":"Beautifiers#getOptionsForPath","p":"class/Beautifiers.html#getOptionsForPath-dynamic"},{"t":"src/beautify.coffee~getScrollTop","p":"file/src/beautify.coffee.html#getScrollTop-"},{"t":"Beautifier#getShellEnvironment","p":"class/Beautifier.html#getShellEnvironment-dynamic"},{"t":"Beautifiers#getUserHome","p":"class/Beautifiers.html#getUserHome-dynamic"},{"t":"LoadingView#hide","p":"class/LoadingView.html#hide-dynamic"},{"t":"Beautifiers#isNestedOptions","p":"class/Beautifiers.html#isNestedOptions-dynamic"},{"t":"Beautifier#readFile","p":"class/Beautifier.html#readFile-dynamic"},{"t":"MessageView#refresh","p":"class/MessageView.html#refresh-dynamic"},{"t":"Beautifier#run","p":"class/Beautifier.html#run-dynamic"},{"t":"src/beautify.coffee~setCursors","p":"file/src/beautify.coffee.html#setCursors-"},{"t":"src/beautify.coffee~setScrollTop","p":"file/src/beautify.coffee.html#setScrollTop-"},{"t":"LatexBeautify#setUpDir","p":"class/LatexBeautify.html#setUpDir-dynamic"},{"t":"Beautifier#setupLogger","p":"class/Beautifier.html#setupLogger-dynamic"},{"t":"MessageView#show","p":"class/MessageView.html#show-dynamic"},{"t":"LoadingView#show","p":"class/LoadingView.html#show-dynamic"},{"t":"src/beautify.coffee~showError","p":"file/src/beautify.coffee.html#showError-"},{"t":"Beautifier#spawn","p":"class/Beautifier.html#spawn-dynamic"},{"t":"Beautifier#tempFile","p":"class/Beautifier.html#tempFile-dynamic"},{"t":"Beautifiers#track","p":"class/Beautifiers.html#track-dynamic"},{"t":"Beautifiers#trackEvent","p":"class/Beautifiers.html#trackEvent-dynamic"},{"t":"Beautifiers#trackTiming","p":"class/Beautifiers.html#trackTiming-dynamic"},{"t":"Beautifiers#transformOptions","p":"class/Beautifiers.html#transformOptions-dynamic"},{"t":"Beautifiers#verifyExists","p":"class/Beautifiers.html#verifyExists-dynamic"},{"t":"Beautifier#which","p":"class/Beautifier.html#which-dynamic"}]
+window.searchData = [{"t":"Autopep8","p":"class/Autopep8.html"},{"t":"Beautifier","p":"class/Beautifier.html"},{"t":"ClangFormat","p":"class/ClangFormat.html"},{"t":"CoffeeFmt","p":"class/CoffeeFmt.html"},{"t":"CoffeeFormatter","p":"class/CoffeeFormatter.html"},{"t":"Crystal","p":"class/Crystal.html"},{"t":"JSBeautify","p":"class/JSBeautify.html"},{"t":"Dfmt","p":"class/Dfmt.html"},{"t":"ElmFormat","p":"class/ElmFormat.html"},{"t":"ErlTidy","p":"class/ErlTidy.html"},{"t":"R","p":"class/R.html"},{"t":"FortranBeautifier","p":"class/FortranBeautifier.html"},{"t":"Gherkin","p":"class/Gherkin.html"},{"t":"Gofmt","p":"class/Gofmt.html"},{"t":"HTMLBeautifier","p":"class/HTMLBeautifier.html"},{"t":"Beautifiers","p":"class/Beautifiers.html"},{"t":"JSBeautify","p":"class/JSBeautify.html"},{"t":"JSCSFixer","p":"class/JSCSFixer.html"},{"t":"LatexBeautify","p":"class/LatexBeautify.html"},{"t":"Lua","p":"class/Lua.html"},{"t":"MarkoBeautifier","p":"class/MarkoBeautifier.html"},{"t":"OCPIndent","p":"class/OCPIndent.html"},{"t":"PerlTidy","p":"class/PerlTidy.html"},{"t":"PHPCSFixer","p":"class/PHPCSFixer.html"},{"t":"PHPCBF","p":"class/PHPCBF.html"},{"t":"PrettyDiff","p":"class/PrettyDiff.html"},{"t":"PugBeautify","p":"class/PugBeautify.html"},{"t":"PuppetFix","p":"class/PuppetFix.html"},{"t":"Remark","p":"class/Remark.html"},{"t":"Rubocop","p":"class/Rubocop.html"},{"t":"RubyBeautify","p":"class/RubyBeautify.html"},{"t":"Rustfmt","p":"class/Rustfmt.html"},{"t":"SassConvert","p":"class/SassConvert.html"},{"t":"Sqlformat","p":"class/Sqlformat.html"},{"t":"StylishHaskell","p":"class/StylishHaskell.html"},{"t":"TidyMarkdown","p":"class/TidyMarkdown.html"},{"t":"TypeScriptFormatter","p":"class/TypeScriptFormatter.html"},{"t":"Uncrustify","p":"class/Uncrustify.html"},{"t":"VueBeautifier","p":"class/VueBeautifier.html"},{"t":"Yapf","p":"class/Yapf.html"},{"t":"LoadingView","p":"class/LoadingView.html"},{"t":"MessageView","p":"class/MessageView.html"},{"t":"src/beautifiers/autopep8.coffee","p":"file/src/beautifiers/autopep8.coffee.html"},{"t":"src/beautifiers/beautifier.coffee","p":"file/src/beautifiers/beautifier.coffee.html"},{"t":"src/beautifiers/clang-format.coffee","p":"file/src/beautifiers/clang-format.coffee.html"},{"t":"src/beautifiers/coffee-fmt.coffee","p":"file/src/beautifiers/coffee-fmt.coffee.html"},{"t":"src/beautifiers/coffee-formatter.coffee","p":"file/src/beautifiers/coffee-formatter.coffee.html"},{"t":"src/beautifiers/crystal.coffee","p":"file/src/beautifiers/crystal.coffee.html"},{"t":"src/beautifiers/csscomb.coffee","p":"file/src/beautifiers/csscomb.coffee.html"},{"t":"src/beautifiers/dfmt.coffee","p":"file/src/beautifiers/dfmt.coffee.html"},{"t":"src/beautifiers/elm-format.coffee","p":"file/src/beautifiers/elm-format.coffee.html"},{"t":"src/beautifiers/erl_tidy.coffee","p":"file/src/beautifiers/erl_tidy.coffee.html"},{"t":"src/beautifiers/formatR/index.coffee","p":"file/src/beautifiers/formatR/index.coffee.html"},{"t":"src/beautifiers/fortran-beautifier/index.coffee","p":"file/src/beautifiers/fortran-beautifier/index.coffee.html"},{"t":"src/beautifiers/gherkin.coffee","p":"file/src/beautifiers/gherkin.coffee.html"},{"t":"src/beautifiers/gofmt.coffee","p":"file/src/beautifiers/gofmt.coffee.html"},{"t":"src/beautifiers/htmlbeautifier.coffee","p":"file/src/beautifiers/htmlbeautifier.coffee.html"},{"t":"src/beautifiers/index.coffee","p":"file/src/beautifiers/index.coffee.html"},{"t":"src/beautifiers/js-beautify.coffee","p":"file/src/beautifiers/js-beautify.coffee.html"},{"t":"src/beautifiers/jscs.coffee","p":"file/src/beautifiers/jscs.coffee.html"},{"t":"src/beautifiers/latex-beautify.coffee","p":"file/src/beautifiers/latex-beautify.coffee.html"},{"t":"src/beautifiers/lua-beautifier/index.coffee","p":"file/src/beautifiers/lua-beautifier/index.coffee.html"},{"t":"src/beautifiers/marko-beautifier.coffee","p":"file/src/beautifiers/marko-beautifier.coffee.html"},{"t":"src/beautifiers/ocp-indent.coffee","p":"file/src/beautifiers/ocp-indent.coffee.html"},{"t":"src/beautifiers/perltidy.coffee","p":"file/src/beautifiers/perltidy.coffee.html"},{"t":"src/beautifiers/php-cs-fixer.coffee","p":"file/src/beautifiers/php-cs-fixer.coffee.html"},{"t":"src/beautifiers/phpcbf.coffee","p":"file/src/beautifiers/phpcbf.coffee.html"},{"t":"src/beautifiers/prettydiff.coffee","p":"file/src/beautifiers/prettydiff.coffee.html"},{"t":"src/beautifiers/pug-beautify.coffee","p":"file/src/beautifiers/pug-beautify.coffee.html"},{"t":"src/beautifiers/puppet-fix.coffee","p":"file/src/beautifiers/puppet-fix.coffee.html"},{"t":"src/beautifiers/remark.coffee","p":"file/src/beautifiers/remark.coffee.html"},{"t":"src/beautifiers/rubocop.coffee","p":"file/src/beautifiers/rubocop.coffee.html"},{"t":"src/beautifiers/ruby-beautify.coffee","p":"file/src/beautifiers/ruby-beautify.coffee.html"},{"t":"src/beautifiers/rustfmt.coffee","p":"file/src/beautifiers/rustfmt.coffee.html"},{"t":"src/beautifiers/sass-convert.coffee","p":"file/src/beautifiers/sass-convert.coffee.html"},{"t":"src/beautifiers/sqlformat.coffee","p":"file/src/beautifiers/sqlformat.coffee.html"},{"t":"src/beautifiers/stylish-haskell.coffee","p":"file/src/beautifiers/stylish-haskell.coffee.html"},{"t":"src/beautifiers/tidy-markdown.coffee","p":"file/src/beautifiers/tidy-markdown.coffee.html"},{"t":"src/beautifiers/typescript-formatter.coffee","p":"file/src/beautifiers/typescript-formatter.coffee.html"},{"t":"src/beautifiers/uncrustify/cfg.coffee","p":"file/src/beautifiers/uncrustify/cfg.coffee.html"},{"t":"src/beautifiers/uncrustify/index.coffee","p":"file/src/beautifiers/uncrustify/index.coffee.html"},{"t":"src/beautifiers/vue-beautifier.coffee","p":"file/src/beautifiers/vue-beautifier.coffee.html"},{"t":"src/beautifiers/yapf.coffee","p":"file/src/beautifiers/yapf.coffee.html"},{"t":"src/beautify.coffee","p":"file/src/beautify.coffee.html"},{"t":"src/config.coffee","p":"file/src/config.coffee.html"},{"t":"src/languages/apex.coffee","p":"file/src/languages/apex.coffee.html"},{"t":"src/languages/arduino.coffee","p":"file/src/languages/arduino.coffee.html"},{"t":"src/languages/c-sharp.coffee","p":"file/src/languages/c-sharp.coffee.html"},{"t":"src/languages/c.coffee","p":"file/src/languages/c.coffee.html"},{"t":"src/languages/coffeescript.coffee","p":"file/src/languages/coffeescript.coffee.html"},{"t":"src/languages/coldfusion.coffee","p":"file/src/languages/coldfusion.coffee.html"},{"t":"src/languages/cpp.coffee","p":"file/src/languages/cpp.coffee.html"},{"t":"src/languages/crystal.coffee","p":"file/src/languages/crystal.coffee.html"},{"t":"src/languages/css.coffee","p":"file/src/languages/css.coffee.html"},{"t":"src/languages/csv.coffee","p":"file/src/languages/csv.coffee.html"},{"t":"src/languages/d.coffee","p":"file/src/languages/d.coffee.html"},{"t":"src/languages/ejs.coffee","p":"file/src/languages/ejs.coffee.html"},{"t":"src/languages/elm.coffee","p":"file/src/languages/elm.coffee.html"},{"t":"src/languages/erb.coffee","p":"file/src/languages/erb.coffee.html"},{"t":"src/languages/erlang.coffee","p":"file/src/languages/erlang.coffee.html"},{"t":"src/languages/fortran.coffee","p":"file/src/languages/fortran.coffee.html"},{"t":"src/languages/gherkin.coffee","p":"file/src/languages/gherkin.coffee.html"},{"t":"src/languages/go.coffee","p":"file/src/languages/go.coffee.html"},{"t":"src/languages/handlebars.coffee","p":"file/src/languages/handlebars.coffee.html"},{"t":"src/languages/haskell.coffee","p":"file/src/languages/haskell.coffee.html"},{"t":"src/languages/html.coffee","p":"file/src/languages/html.coffee.html"},{"t":"src/languages/index.coffee","p":"file/src/languages/index.coffee.html"},{"t":"src/languages/jade.coffee","p":"file/src/languages/jade.coffee.html"},{"t":"src/languages/java.coffee","p":"file/src/languages/java.coffee.html"},{"t":"src/languages/javascript.coffee","p":"file/src/languages/javascript.coffee.html"},{"t":"src/languages/json.coffee","p":"file/src/languages/json.coffee.html"},{"t":"src/languages/jsx.coffee","p":"file/src/languages/jsx.coffee.html"},{"t":"src/languages/latex.coffee","p":"file/src/languages/latex.coffee.html"},{"t":"src/languages/less.coffee","p":"file/src/languages/less.coffee.html"},{"t":"src/languages/lua.coffee","p":"file/src/languages/lua.coffee.html"},{"t":"src/languages/markdown.coffee","p":"file/src/languages/markdown.coffee.html"},{"t":"src/languages/marko.coffee","p":"file/src/languages/marko.coffee.html"},{"t":"src/languages/mustache.coffee","p":"file/src/languages/mustache.coffee.html"},{"t":"src/languages/nunjucks.coffee","p":"file/src/languages/nunjucks.coffee.html"},{"t":"src/languages/objective-c.coffee","p":"file/src/languages/objective-c.coffee.html"},{"t":"src/languages/ocaml.coffee","p":"file/src/languages/ocaml.coffee.html"},{"t":"src/languages/pawn.coffee","p":"file/src/languages/pawn.coffee.html"},{"t":"src/languages/perl.coffee","p":"file/src/languages/perl.coffee.html"},{"t":"src/languages/php.coffee","p":"file/src/languages/php.coffee.html"},{"t":"src/languages/puppet.coffee","p":"file/src/languages/puppet.coffee.html"},{"t":"src/languages/python.coffee","p":"file/src/languages/python.coffee.html"},{"t":"src/languages/r.coffee","p":"file/src/languages/r.coffee.html"},{"t":"src/languages/riotjs.coffee","p":"file/src/languages/riotjs.coffee.html"},{"t":"src/languages/ruby.coffee","p":"file/src/languages/ruby.coffee.html"},{"t":"src/languages/rust.coffee","p":"file/src/languages/rust.coffee.html"},{"t":"src/languages/sass.coffee","p":"file/src/languages/sass.coffee.html"},{"t":"src/languages/scss.coffee","p":"file/src/languages/scss.coffee.html"},{"t":"src/languages/spacebars.coffee","p":"file/src/languages/spacebars.coffee.html"},{"t":"src/languages/sql.coffee","p":"file/src/languages/sql.coffee.html"},{"t":"src/languages/svg.coffee","p":"file/src/languages/svg.coffee.html"},{"t":"src/languages/swig.coffee","p":"file/src/languages/swig.coffee.html"},{"t":"src/languages/tss.coffee","p":"file/src/languages/tss.coffee.html"},{"t":"src/languages/twig.coffee","p":"file/src/languages/twig.coffee.html"},{"t":"src/languages/typescript.coffee","p":"file/src/languages/typescript.coffee.html"},{"t":"src/languages/vala.coffee","p":"file/src/languages/vala.coffee.html"},{"t":"src/languages/visualforce.coffee","p":"file/src/languages/visualforce.coffee.html"},{"t":"src/languages/vue.coffee","p":"file/src/languages/vue.coffee.html"},{"t":"src/languages/xml.coffee","p":"file/src/languages/xml.coffee.html"},{"t":"src/languages/xtemplate.coffee","p":"file/src/languages/xtemplate.coffee.html"},{"t":"src/views/loading-view.coffee","p":"file/src/views/loading-view.coffee.html"},{"t":"src/views/message-view.coffee","p":"file/src/views/message-view.coffee.html"},{"t":"README.md","p":"extra/README.md.html"},{"t":"CHANGELOG.md","p":"extra/CHANGELOG.md.html"},{"t":"MessageView#addMessage","p":"class/MessageView.html#addMessage-dynamic"},{"t":"Gofmt#beautify","p":"class/Gofmt.html#beautify-dynamic"},{"t":"OCPIndent#beautify","p":"class/OCPIndent.html#beautify-dynamic"},{"t":"src/beautify.coffee~beautify","p":"file/src/beautify.coffee.html#beautify-"},{"t":"MarkoBeautifier#beautify","p":"class/MarkoBeautifier.html#beautify-dynamic"},{"t":"Lua#beautify","p":"class/Lua.html#beautify-dynamic"},{"t":"LatexBeautify#beautify","p":"class/LatexBeautify.html#beautify-dynamic"},{"t":"Rustfmt#beautify","p":"class/Rustfmt.html#beautify-dynamic"},{"t":"Autopep8#beautify","p":"class/Autopep8.html#beautify-dynamic"},{"t":"JSCSFixer#beautify","p":"class/JSCSFixer.html#beautify-dynamic"},{"t":"RubyBeautify#beautify","p":"class/RubyBeautify.html#beautify-dynamic"},{"t":"JSBeautify#beautify","p":"class/JSBeautify.html#beautify-dynamic"},{"t":"PHPCSFixer#beautify","p":"class/PHPCSFixer.html#beautify-dynamic"},{"t":"Rubocop#beautify","p":"class/Rubocop.html#beautify-dynamic"},{"t":"SassConvert#beautify","p":"class/SassConvert.html#beautify-dynamic"},{"t":"PHPCBF#beautify","p":"class/PHPCBF.html#beautify-dynamic"},{"t":"PrettyDiff#beautify","p":"class/PrettyDiff.html#beautify-dynamic"},{"t":"PugBeautify#beautify","p":"class/PugBeautify.html#beautify-dynamic"},{"t":"Yapf#beautify","p":"class/Yapf.html#beautify-dynamic"},{"t":"Sqlformat#beautify","p":"class/Sqlformat.html#beautify-dynamic"},{"t":"StylishHaskell#beautify","p":"class/StylishHaskell.html#beautify-dynamic"},{"t":"ClangFormat#beautify","p":"class/ClangFormat.html#beautify-dynamic"},{"t":"CoffeeFmt#beautify","p":"class/CoffeeFmt.html#beautify-dynamic"},{"t":"CoffeeFormatter#beautify","p":"class/CoffeeFormatter.html#beautify-dynamic"},{"t":"Crystal#beautify","p":"class/Crystal.html#beautify-dynamic"},{"t":"JSBeautify#beautify","p":"class/JSBeautify.html#beautify-dynamic"},{"t":"Dfmt#beautify","p":"class/Dfmt.html#beautify-dynamic"},{"t":"ElmFormat#beautify","p":"class/ElmFormat.html#beautify-dynamic"},{"t":"ErlTidy#beautify","p":"class/ErlTidy.html#beautify-dynamic"},{"t":"R#beautify","p":"class/R.html#beautify-dynamic"},{"t":"FortranBeautifier#beautify","p":"class/FortranBeautifier.html#beautify-dynamic"},{"t":"Gherkin#beautify","p":"class/Gherkin.html#beautify-dynamic"},{"t":"Beautifiers#beautify","p":"class/Beautifiers.html#beautify-dynamic"},{"t":"HTMLBeautifier#beautify","p":"class/HTMLBeautifier.html#beautify-dynamic"},{"t":"TidyMarkdown#beautify","p":"class/TidyMarkdown.html#beautify-dynamic"},{"t":"TypeScriptFormatter#beautify","p":"class/TypeScriptFormatter.html#beautify-dynamic"},{"t":"Remark#beautify","p":"class/Remark.html#beautify-dynamic"},{"t":"Uncrustify#beautify","p":"class/Uncrustify.html#beautify-dynamic"},{"t":"PuppetFix#beautify","p":"class/PuppetFix.html#beautify-dynamic"},{"t":"VueBeautifier#beautify","p":"class/VueBeautifier.html#beautify-dynamic"},{"t":"PerlTidy#beautify","p":"class/PerlTidy.html#beautify-dynamic"},{"t":"src/beautify.coffee~beautifyDirectory","p":"file/src/beautify.coffee.html#beautifyDirectory-"},{"t":"src/beautify.coffee~beautifyFile","p":"file/src/beautify.coffee.html#beautifyFile-"},{"t":"src/beautify.coffee~beautifyFilePath","p":"file/src/beautify.coffee.html#beautifyFilePath-"},{"t":"LatexBeautify#buildConfigFile","p":"class/LatexBeautify.html#buildConfigFile-dynamic"},{"t":"MessageView#clearMessages","p":"class/MessageView.html#clearMessages-dynamic"},{"t":"PerlTidy#cli","p":"class/PerlTidy.html#cli-dynamic"},{"t":"PuppetFix#cli","p":"class/PuppetFix.html#cli-dynamic"},{"t":"MessageView#close","p":"class/MessageView.html#close-dynamic"},{"t":"Beautifier#commandNotFoundError","p":"class/Beautifier.html#commandNotFoundError-dynamic"},{"t":"MessageView#constructor","p":"class/MessageView.html#constructor-dynamic"},{"t":"Beautifier#constructor","p":"class/Beautifier.html#constructor-dynamic"},{"t":"Beautifiers#constructor","p":"class/Beautifiers.html#constructor-dynamic"},{"t":"MessageView.content","p":"class/MessageView.html#content-static"},{"t":"LoadingView.content","p":"class/LoadingView.html#content-static"},{"t":"src/beautify.coffee~debug","p":"file/src/beautify.coffee.html#debug-"},{"t":"Beautifier#deprecate","p":"class/Beautifier.html#deprecate-dynamic"},{"t":"MessageView#destroy","p":"class/MessageView.html#destroy-dynamic"},{"t":"ClangFormat#dumpToFile","p":"class/ClangFormat.html#dumpToFile-dynamic"},{"t":"Beautifiers#encodeURI","p":"class/Beautifiers.html#encodeURI-dynamic"},{"t":"Beautifiers#findConfig","p":"class/Beautifiers.html#findConfig-dynamic"},{"t":"Beautifier#findFile","p":"class/Beautifier.html#findFile-dynamic"},{"t":"Beautifiers#findFile","p":"class/Beautifiers.html#findFile-dynamic"},{"t":"Beautifiers#getBeautifierForLanguage","p":"class/Beautifiers.html#getBeautifierForLanguage-dynamic"},{"t":"Beautifiers#getBeautifiers","p":"class/Beautifiers.html#getBeautifiers-dynamic"},{"t":"Beautifiers#getConfig","p":"class/Beautifiers.html#getConfig-dynamic"},{"t":"Beautifiers#getConfigOptionsFromSettings","p":"class/Beautifiers.html#getConfigOptionsFromSettings-dynamic"},{"t":"src/beautify.coffee~getCursors","p":"file/src/beautify.coffee.html#getCursors-"},{"t":"Beautifiers#getLanguage","p":"class/Beautifiers.html#getLanguage-dynamic"},{"t":"Beautifiers#getOptions","p":"class/Beautifiers.html#getOptions-dynamic"},{"t":"Beautifiers#getOptionsForLanguage","p":"class/Beautifiers.html#getOptionsForLanguage-dynamic"},{"t":"Beautifiers#getOptionsForPath","p":"class/Beautifiers.html#getOptionsForPath-dynamic"},{"t":"src/beautify.coffee~getScrollTop","p":"file/src/beautify.coffee.html#getScrollTop-"},{"t":"Beautifier#getShellEnvironment","p":"class/Beautifier.html#getShellEnvironment-dynamic"},{"t":"Beautifiers#getUserHome","p":"class/Beautifiers.html#getUserHome-dynamic"},{"t":"LoadingView#hide","p":"class/LoadingView.html#hide-dynamic"},{"t":"Beautifiers#isNestedOptions","p":"class/Beautifiers.html#isNestedOptions-dynamic"},{"t":"Beautifier#readFile","p":"class/Beautifier.html#readFile-dynamic"},{"t":"MessageView#refresh","p":"class/MessageView.html#refresh-dynamic"},{"t":"Beautifier#run","p":"class/Beautifier.html#run-dynamic"},{"t":"src/beautify.coffee~setCursors","p":"file/src/beautify.coffee.html#setCursors-"},{"t":"src/beautify.coffee~setScrollTop","p":"file/src/beautify.coffee.html#setScrollTop-"},{"t":"LatexBeautify#setUpDir","p":"class/LatexBeautify.html#setUpDir-dynamic"},{"t":"Beautifier#setupLogger","p":"class/Beautifier.html#setupLogger-dynamic"},{"t":"MessageView#show","p":"class/MessageView.html#show-dynamic"},{"t":"LoadingView#show","p":"class/LoadingView.html#show-dynamic"},{"t":"src/beautify.coffee~showError","p":"file/src/beautify.coffee.html#showError-"},{"t":"Beautifier#spawn","p":"class/Beautifier.html#spawn-dynamic"},{"t":"Beautifier#tempFile","p":"class/Beautifier.html#tempFile-dynamic"},{"t":"Beautifiers#track","p":"class/Beautifiers.html#track-dynamic"},{"t":"Beautifiers#trackEvent","p":"class/Beautifiers.html#trackEvent-dynamic"},{"t":"Beautifiers#trackTiming","p":"class/Beautifiers.html#trackTiming-dynamic"},{"t":"Beautifiers#transformOptions","p":"class/Beautifiers.html#transformOptions-dynamic"},{"t":"Beautifiers#verifyExists","p":"class/Beautifiers.html#verifyExists-dynamic"},{"t":"Beautifier#which","p":"class/Beautifier.html#which-dynamic"}]
