@@ -3,25 +3,24 @@
 import _ from 'underscore-plus';
 
 export default class Settings {
-  update(settings={}) {
+  update(settings = {}) {
     this.load(settings);
   }
 
-  load(settings={}) {
+  load(values = {}) {
+    let settings = values;
     if ('global' in settings) {
       settings['*'] = settings.global;
       delete settings.global;
     }
 
     if ('*' in settings) {
-      let scopedSettings = settings;
+      const scopedSettings = settings;
       settings = settings['*'];
       delete scopedSettings['*'];
 
-      let setting;
-      let scope;
-      for (scope in scopedSettings) {
-        setting = scopedSettings[scope];
+      for (const scope of Object.keys(scopedSettings)) {
+        const setting = scopedSettings[scope];
         this.set(setting, scope);
       }
     }
@@ -30,28 +29,24 @@ export default class Settings {
   }
 
   set(settings, scope) {
-    let flatSettings = {};
-    let setting;
+    const flatSettings = {};
+    const options = scope ? { scopeSelector: scope } : {};
     let value;
-    let valueOptions;
-    let currentValue;
-    let options = scope ? {scopeSelector: scope} : {};
     options.save = false;
     this.flatten(flatSettings, settings);
 
-    for (setting in flatSettings) {
-      value = flatSettings[setting];
-
-      atom.config.set(setting, value, options);
+    for (const key of Object.keys(flatSettings)) {
+      value = flatSettings[key];
+      atom.config.set(key, value, options);
     }
   }
 
   flatten(root, dict, path) {
-    let key;
     let value;
     let dotPath;
     let isObject;
-    for (key in dict) {
+
+    for (const key of Object.keys(dict)) {
       value = dict[key];
       dotPath = path ? `${path}.${key}` : key;
       isObject = !_.isArray(value) && _.isObject(value);
